zigbee-simulator/requirements.txt
# /home/andrey/Projects/garage/zigbee-simulator/requirements.txt
paho-mqtt==1.6.1
